refactor(MovieDetail): 디렉터리 위치 변경
